requirements.txt
python-dotenv
openai
gradio
anthropic
pandas